+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.20] - 2025-04-09
+
+### Added
+
+- Add jpg to supported export format
+
+### Fixed
+
+- Fix Get-DiaNodeIcon cmdlet to honor align parameter
+
 ## [0.2.19] - 2025-03-04
 
 ### Added